requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
pytest==8.3.3
pytest-cov==5.0.0
